--- Leetcode/October_Leetcode_Challenge/Day6_Insert_into_a_Binary_Search_Tree.cpp
@@ -1,33 +1,0 @@
-/**
- * Definition for a binary tree node.
- * struct TreeNode {
- *     int val;
- *     TreeNode *left;
- *     TreeNode *right;
- *     TreeNode() : val(0), left(nullptr), right(nullptr) {}
- *     TreeNode(int x) : val(x), left(nullptr), right(nullptr) {}
- *     TreeNode(int x, TreeNode *left, TreeNode *right) : val(x), left(left), right(right) {}
- * };
- */
-
-class Solution
-{
-public:
-    TreeNode *insertIntoBST(TreeNode *root, int val)
-    {
-
-        if (root == NULL)
-        {
-            return new TreeNode(val);
-        }
-        else if (root->val > val)
-        {
-            root->left = insertIntoBST(root->left, val);
-        }
-        else if (root->val < val)
-        {
-            root->right = insertIntoBST(root->right, val);
-        }
-        return root;
-    }
-};
